Updated the example to cater for sequences of lists. That is the list in the events bin is now bounded in size controlled by the MAX_RECORDS_IN_SEQUENCE value. Once more than this number of records is attempted to be written, a new key will be added with a sequence number dictated by the number of records.
For example, assume MAX_RECORDS_IN_SEQUENCE is 100, and 100 records for User-125 have
been written on 09-08-2015. There will be a key in the database with:
User-125:09-08-2015:1

However, when the next record is written an additional records will be in the database,
so there will now be:
User-125:09-08-2015:1
User-125-09-08-2015:2

Hence, no single entry in the database will store more than MAX_RECORDS_IN_SEQUENCE values.